pyproject.toml
[project]
name = "nonebot_plugin_impact"
version = "0.07.114514" 
description = "让群友们眼前一黑的nonebot2淫趴插件"
authors = [
    {name = "Special-Week", email = "HuaMing27499@gmail.com"},
]
dependencies = [
    "pillow>=9.1.1",
    "nonebot2>=2.0.0b5",
    "nonebot-adapter-onebot>=2.2.3",
]
requires-python = ">=3.8.0"
readme = "README.md"
license = {text = "MIT"}